+++ packages/next-common/components/postList/post.jsx
@@ -14,7 +14,7 @@ import { getGov2ReferendumStateArgs } from "../../utils/gov2/result";
 import { useChain, useChainSettings } from "../../context/chain";
 import Gov2TrackTag from "../gov2/trackTag";
 import DecisionCountdown from "../gov2/postList/decisionCountdown";
-import { gov2State, gov2VotingState } from "../../utils/consts/state";
+import { gov2State, gov2VotingStates } from "../../utils/consts/state";
 import ConfirmCountdown from "../gov2/postList/confirmCountdown";
 import ValueDisplay from "../valueDisplay";
 import ListPostTitle from "./postTitle";
@@ -33,19 +33,17 @@ import PolkassemblyUser from "../user/polkassemblyUser";
 import Tooltip from "next-common/components/tooltip";
 import WarningIcon from "next-common/assets/imgs/icons/warning.svg";
 import { getAssetByMeta } from "next-common/utils/treasury/spend/usdCheck";
-import { SystemActivity, SystemComment } from "@osn/icons/subsquare";
+import { SystemComment } from "@osn/icons/subsquare";
 import PostListTreasuryAllSpends from "./treasuryAllSpends";
-import { formatTimeAgo } from "next-common/utils/viewfuncs/formatTimeAgo";
 import PostListAISummary from "./aiSummary";
-import TreasurySpendsCountDown from "next-common/components/postList/treasury/spends/countdown";
 import PostListMyVoteMark from "./myVoteMark";
 import { referendumState } from "next-common/utils/consts/referendum";
 import Chains from "next-common/utils/consts/chains";
+import { PostItemTime } from "./common";
 
 import {
   Wrapper,
   Footer,
-  Info,
   MobileHiddenInfo,
   FooterWrapper,
   TitleExtraValue,
@@ -122,12 +120,7 @@ export function PostValueTitle({ data, type }) {
     ? onchainData?.treasuryInfo?.amount
     : value;
 
-  if (
-    [
-      businessCategory.treasurySpends,
-      businessCategory.fellowshipTreasurySpends,
-    ].includes(type)
-  ) {
+  if ([businessCategory.fellowshipTreasurySpends].includes(type)) {
     return <TreasurySpendAmount meta={data?.meta} />;
   }
 
@@ -174,11 +167,9 @@ export default function Post({ data, href, type }) {
     businessCategory.openTechCommitteeProposals,
   ].includes(type);
 
-  const isGov2Referendum = [
-    businessCategory.openGovReferenda,
-    businessCategory.fellowship,
-    businessCategory.ambassadorReferenda,
-  ].includes(type);
+  const isGov2Referendum = [businessCategory.ambassadorReferenda].includes(
+    type,
+  );
 
   let stateArgs;
   if (isDemocracyCollective) {
@@ -191,9 +182,6 @@ export default function Post({ data, href, type }) {
       data.onchainData.timeline,
     );
   }
-
-  const timeAgo = formatTimeAgo(data.time);
-  const createAgo = formatTimeAgo(data.createdAt);
 
   let elapseIcon = null;
   if (
@@ -212,12 +200,7 @@ export default function Post({ data, href, type }) {
 
   if (isGov2Referendum) {
     if (data?.status === gov2State.Preparing) {
-      elapseIcon = (
-        <PreparingCountdown
-          detail={data}
-          isFellowship={businessCategory.fellowship === type}
-        />
-      );
+      elapseIcon = <PreparingCountdown detail={data} />;
     } else if (data?.status === gov2State.Deciding) {
       elapseIcon = <DecisionCountdown detail={data} />;
     } else if (data?.status === gov2State.Confirming) {
@@ -227,13 +210,13 @@ export default function Post({ data, href, type }) {
 
   if (businessCategory.democracyReferenda === type) {
     elapseIcon = <ReferendumElapse detail={data} />;
-  } else if (businessCategory.treasurySpends === type) {
-    elapseIcon = <TreasurySpendsCountDown data={data} />;
   }
 
   let commentsCount = data.commentsCount || 0;
   if (
-    [Chains.kusama, Chains.kusamaPeople].includes(currentChain) &&
+    [Chains.kusama, Chains.kusamaPeople, Chains.polkadot].includes(
+      currentChain,
+    ) &&
     data.polkassemblyCommentsCount
   ) {
     commentsCount = data.polkassemblyCommentsCount || 0;
@@ -242,24 +225,18 @@ export default function Post({ data, href, type }) {
   const bannerUrl = getBannerUrl(data.bannerCid);
 
   let trackTagLink = null;
-  if (type === businessCategory.openGovReferenda) {
-    trackTagLink = `/referenda/tracks/${data.track}`;
-  } else if (type === businessCategory.fellowship) {
-    trackTagLink = `/fellowship/tracks/${data.track}`;
-  } else if (type === businessCategory.ambassadorReferenda) {
+  if (type === businessCategory.ambassadorReferenda) {
     trackTagLink = `/ambassador/tracks/${data.track}`;
   }
 
   const hasTally = data.onchainData?.tally || data.onchainData?.info?.tally;
   const showTally = [
     businessCategory.democracyReferenda,
-    businessCategory.openGovReferenda,
-    businessCategory.fellowship,
     businessCategory.ambassadorReferenda,
   ].includes(type);
 
   const showVoteMark =
-    (isGov2Referendum && gov2VotingState.includes(data?.status)) ||
+    (isGov2Referendum && gov2VotingStates.includes(data?.status)) ||
     (businessCategory.democracyReferenda === type &&
       data?.status === referendumState.Started);
 
@@ -296,25 +273,7 @@ export default function Post({ data, href, type }) {
                 <DemocracyTag />
               </div>
             )}
-            {data.time && (
-              <Info>
-                <SystemActivity className="w-4 h-4 stroke-textTertiary [&_path]:stroke-2" />
-                <Tooltip
-                  className="flex"
-                  content={
-                    <div className="text12Medium">
-                      <ul className="list-disc list-inside">
-                        <li>Created at {createAgo}</li>
-                        <li>Latest activity at {timeAgo}</li>
-                      </ul>
-                    </div>
-                  }
-                >
-                  <span className="cursor-pointer">{timeAgo}</span>
-                </Tooltip>
-                <Flex className="elapseIcon">{elapseIcon}</Flex>
-              </Info>
-            )}
+            <PostItemTime data={data} elapseIcon={elapseIcon} />
             {commentsCount > -1 && (
               <MobileHiddenInfo>
                 <Tooltip