+++ packages/next-common/components/summary/newProposalQuickStart/common/useAutoSelectTreasuryTrackField.js
@@ -2,9 +2,13 @@ import { useState } from "react";
 import { usePageProps } from "next-common/context/page";
 import AutoSelectTreasuryTrack from "next-common/components/popup/fields/autoSelectTreasuryTrack";
 
-export default function useAutoSelectTreasuryTrackField(requestAmount) {
+export default function useAutoSelectTreasuryTrackField(
+  requestAmount,
+  defaultValue = null,
+) {
   const { tracks } = usePageProps();
-  const [trackId, setTrackId] = useState(tracks[0].id);
+  const defaultTrackId = defaultValue || tracks[0].id;
+  const [trackId, setTrackId] = useState(defaultTrackId);
 
   return {
     value: trackId,