+++ packages/next-common/context/relationship/index.js
@@ -19,7 +19,6 @@ const defaultContext = {
 const RelationshipContext = createContext(defaultContext);
 
 export default function RelationshipProvider({
-  rootAddress,
   children,
   nodes,
   edges,
@@ -30,21 +29,13 @@ export default function RelationshipProvider({
   );
   const value = useMemo(
     () => ({
-      rootAddress,
       nodes,
       edges,
       isLoading,
       excludedIndications,
       setExcludedIndications,
     }),
-    [
-      rootAddress,
-      nodes,
-      edges,
-      isLoading,
-      excludedIndications,
-      setExcludedIndications,
-    ],
+    [nodes, edges, isLoading, excludedIndications, setExcludedIndications],
   );
 
   return (
@@ -148,9 +139,4 @@ export function useSwitchIndications() {
     setExcludedIndications,
     toggleIndication,
   };
-}
-
-export function useRootAddress() {
-  const context = useContext(RelationshipContext);
-  return context?.rootAddress || "";
 }