+++ packages/next/pages/treasury/child-bounties/index.js
@@ -1,4 +1,4 @@
-import PostList from "next-common/components/postList";
+import TreasuryChildBountiesPostList from "next-common/components/postList/treasuryChildBountiesPostList";
 import { withCommonProps } from "next-common/lib";
 import { toTreasuryChildBountyListItem } from "next-common/utils/viewfuncs";
 import { useChain } from "next-common/context/chain";
@@ -40,9 +40,7 @@ export default function ChildBountiesPage({ bounties }) {
           },
         ].filter(Boolean)}
       >
-        <PostList
-          category={category}
-          title="List"
+        <TreasuryChildBountiesPostList
           titleCount={bounties.total}
           items={items}
           pagination={{