+++ packages/next/pages/treasury/proposals/index.js
@@ -1,5 +1,7 @@
 import { useEffect, useState } from "react";
-import PostList from "next-common/components/postList";
+import TreasuryProposalsPostList, {
+  NewTreasuryProposalButton,
+} from "next-common/components/postList/treasuryProposalsPostList";
 import { withCommonProps } from "next-common/lib";
 import TreasurySummary from "next-common/components/summary/treasurySummary";
 import normalizeTreasuryProposalListItem from "next-common/utils/viewfuncs/treasury/normalizeProposalListItem";
@@ -11,20 +13,20 @@ import {
   TreasuryProvider,
   useTreasuryProposalListUrl,
 } from "next-common/context/treasury";
-import NewTreasuryProposal from "next-common/components/treasury/proposal/newTreasuryProposal";
 import { isPolkadotChain } from "next-common/utils/chain";
 import PolkadotTreasuryStatsOnProposal from "next-common/components/treasury/common/polkadotTreasuryStatsOnProposal";
+import businessCategory from "next-common/utils/consts/business/category";
 
 export default function ProposalsPage({ proposals: ssrProposals, chain }) {
   const [proposals, setProposals] = useState(ssrProposals);
   useEffect(() => setProposals(ssrProposals), [ssrProposals]);
-  const { integrations, showNewTreasuryProposalButton } = useChainSettings();
+  const { integrations } = useChainSettings();
 
   const items = (proposals.items || []).map((item) =>
     normalizeTreasuryProposalListItem(chain, item),
   );
 
-  const category = "Treasury Proposals";
+  const category = businessCategory.treasuryProposals;
   const seoInfo = { title: category, desc: category };
 
   const pallet = "treasury";
@@ -55,16 +57,7 @@ export default function ProposalsPage({ proposals: ssrProposals, chain }) {
           },
         ].filter(Boolean)}
       >
-        <PostList
-          titleExtra={
-            showNewTreasuryProposalButton && (
-              <div className="flex justify-end">
-                <NewTreasuryProposal />
-              </div>
-            )
-          }
-          category={category}
-          title="List"
+        <TreasuryProposalsPostList
           titleCount={proposals.total}
           items={items}
           pagination={{
@@ -72,6 +65,7 @@ export default function ProposalsPage({ proposals: ssrProposals, chain }) {
             pageSize: proposals.pageSize,
             total: proposals.total,
           }}
+          titleExtra={<NewTreasuryProposalButton />}
         />
       </ListLayout>
     </TreasuryProvider>